+++ content_migration/management/commands/import_magazine.py
@@ -1,37 +1,15 @@
-from django.core.management import call_command
-from django.core.management.base import BaseCommand
+from django.core.management.base import BaseCommand, CommandParser
+
+from content_migration.management.import_magazine_handler import handle_import_magazine
 
 
 class Command(BaseCommand):
     help = "Import all magazine content"
 
-    def add_arguments(self, parser):
+    def add_arguments(self, parser: CommandParser) -> None:
         parser.add_argument("--data-directory", action="store", type=str)
 
-    def handle(self, *args, **options):
+    def handle(self, *args: tuple, **options: dict[str, str]) -> None:
         directory = options["data_directory"]
 
-        if not directory.endswith("/"):
-            directory += "/"
-
-        magazine_departments_filename = "magazine_departments.csv"
-        magazine_authors_filename = "magazine_authors.csv"
-        magazine_issues_filename = "magazine_issues.csv"
-        magazine_articles_filename = "magazine_articles.csv"
-
-        call_command(
-            "import_magazine_departments",
-            file=f"{ directory }{ magazine_departments_filename }",
-        )
-        call_command(
-            "import_magazine_authors",
-            file=f"{ directory }{ magazine_authors_filename }",
-        )
-        call_command(
-            "import_magazine_issues", file=f"{ directory }{ magazine_issues_filename }"
-        )
-        call_command(
-            "import_magazine_articles",
-            articles_file=f"{ directory }{ magazine_articles_filename }",
-            authors_file=f"{ directory }{ magazine_authors_filename }",
-        )
+        handle_import_magazine(directory)  # type: ignore
