+++ content_migration/management/commands/import_pages.py
@@ -1,40 +1,15 @@
-from django.core.management.base import BaseCommand
-import numpy as np
-import pandas as pd
-from tqdm import tqdm
-from content_migration.management.commands.shared import parse_body_blocks
+from django.core.management.base import BaseCommand, CommandParser
 
-from home.models import HomePage
-from news.models import NewsIndexPage, NewsItem
+from content_migration.management.import_pages_handler import handle_import_pages
 
 
 class Command(BaseCommand):
     help = "Import all Drupal pages"
 
-    def add_arguments(self, parser):
+    def add_arguments(self, parser: CommandParser) -> None:
         parser.add_argument("--file", action="store", type=str)
 
-    def handle(self, *args, **options):
-        # Get references to relevant index pages
-        HomePage.objects.get()
-        news_index_page = NewsIndexPage.objects.get()
+    def handle(self, *args: tuple, **options: dict[str, str]) -> None:
+        file_name = options["file"]
 
-        pages = pd.read_csv(options["file"]).replace({np.nan: None}).to_dict("records")
-
-        for page in tqdm(
-            pages,
-            total=len(pages),
-            desc="Pages",
-            unit="row",
-        ):
-            if page["collection"] == "Extra Extra":
-                news_item = NewsItem(
-                    title=page["title"],
-                    teaser="",
-                    drupal_node_id=page["drupal_node_id"],
-                    body=parse_body_blocks(page["body"]),
-                    body_migrated=page["body"],
-                    publication_date=page["publication_date"],
-                )
-
-                news_index_page.add_child(instance=news_item)
+        handle_import_pages(file_name)  # type: ignore
